+++ protected/humhub/docs/CHANGELOG.md
@@ -4,6 +4,8 @@ HumHub Change Log
 - Fix: Invite error in french language
 - Fix #2518: ActivityStreamWidget::EVENT_INIT is missed (githubjeka)
 - Enh: Fixed accessibility issues in Dashboard/Login/Profile
+- Fix: module beforeInit and afterInit event
+- Enh: Added Registraion::EVENT_AFTER_REGISTRATION UserEvent
 
 1.2.0 (April 16, 2017)
 --------------------------------